requirements.txt
absl-py==0.15.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
beautifulsoup4==4.10.0
braceexpand==0.1.7
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
cycler==0.11.0
DateTime==4.3
flatbuffers==2.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
idna==3.3
imageio==2.10.3
import-expression==1.1.4
jishaku==2.3.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
matplotlib==3.4.3
multidict==5.2.0
networkx==2.6.3
NudeNet==2.0.9
numpy==1.21.4
oauthlib==3.1.1
onnxruntime==1.9.0
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
Pillow==8.4.0
progressbar2==3.55.0
protobuf==3.19.1
psutil==5.8.0
py-cord==1.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydload==1.0.9
pyparsing==3.0.5
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.2
six==1.16.0
smmap==5.0.0
soupsieve==2.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tifffile==2021.11.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zope.interface==5.4.0
